+++ tests/test_merge_modules.py
@@ -2,6 +2,8 @@ import unittest
 
 import torch
 
+from gromo.containers.growing_dag import Expansion
+from gromo.containers.growing_graph_network import GrowingGraphNetwork
 from gromo.modules.conv2d_growing_module import (
     Conv2dGrowingModule,
     Conv2dMergeGrowingModule,
@@ -42,6 +44,7 @@ class TestMergeGrowingModules(unittest.TestCase):
             out_channels=self.out_channels,
             kernel_size=self.kernel_size,
             input_size=self.input_shape,
+            name="conv",
         )
         self.conv_merge = Conv2dMergeGrowingModule(
             in_channels=self.out_channels,
@@ -49,12 +52,16 @@ class TestMergeGrowingModules(unittest.TestCase):
             input_size=(self.out_dim, self.out_dim),
             input_volume=self.linear_in_features,
             post_merge_function=torch.nn.ReLU(),
+            name="conv_merge",
         )
         self.flatten = torch.nn.Flatten()
-        self.linear_merge = LinearMergeGrowingModule(in_features=self.linear_in_features)
+        self.linear_merge = LinearMergeGrowingModule(
+            in_features=self.linear_in_features, name="linear_merge"
+        )
         self.linear = LinearGrowingModule(
             in_features=self.linear_in_features,
             out_features=self.out_features,
+            name="linear",
         )
         self.layers = [self.conv, self.conv_merge, self.linear_merge, self.linear]
 
@@ -157,6 +164,7 @@ class TestMergeGrowingModules(unittest.TestCase):
             in_features=2,
             out_features=self.linear_in_features,
             post_layer_function=torch.nn.ReLU(),
+            name="linear_intercept",
         )
         self.layers.append(linear_intercept)
         x_intercept = torch.rand(self.batch_size, 2, device=self.device)
@@ -256,6 +264,7 @@ class TestMergeGrowingModules(unittest.TestCase):
             in_channels=self.out_channels,
             out_channels=1,
             kernel_size=self.kernel_size,
+            name="conv_early_exit",
             # input_size=self.input_shape,
             # post_layer_function=torch.nn.Sequential(
             #     torch.nn.Flatten(start_dim=1),
@@ -357,6 +366,387 @@ class TestMergeGrowingModules(unittest.TestCase):
 
         # Assert activities
         self.assertTrue(torch.all(self.conv_merge.activity == conv_early_exit.input))
+
+        # Reset computation
+        for layer in self.layers:
+            layer.reset_computation()
+            layer.delete_update()
+
+    def test_conv2d_merge_to_conv2d_merge_with_pooling(self):
+        # Create modules
+        out_dim_after_pooling = (self.out_dim - self.kernel_size[0]) // self.kernel_size[
+            0
+        ] + 1
+        new_linear_in_features = (
+            out_dim_after_pooling * out_dim_after_pooling * self.out_channels
+        )
+        pooling = torch.nn.AvgPool2d(kernel_size=self.kernel_size)
+        conv_merge_after_pooling = Conv2dMergeGrowingModule(
+            in_channels=self.out_channels,
+            next_kernel_size=(1, 1),
+            input_size=(out_dim_after_pooling, out_dim_after_pooling),
+            name="conv2d_merge_after_pooling",
+        )
+        new_linear_merge = LinearMergeGrowingModule(
+            in_features=new_linear_in_features, name="new_linear_merge"
+        )
+        new_linear = LinearGrowingModule(
+            in_features=new_linear_in_features,
+            out_features=self.out_features,
+            name="new_linear",
+        )
+
+        self.layers.remove(self.linear_merge)
+        self.layers.remove(self.linear)
+        self.layers.append(conv_merge_after_pooling)
+        self.layers.append(new_linear_merge)
+        self.layers.append(new_linear)
+
+        # Direct connection between Conv2dMergeGrowingModule and Conv2dMergeGrowingModule
+        self.conv.next_module = self.conv_merge
+        self.conv_merge.add_previous_module(self.conv)
+        self.conv_merge.add_next_module(conv_merge_after_pooling)
+        # self.conv_merge.add_next_module(pooling)
+        # conv_merge_after_pooling.add_previous_module(pooling)
+        conv_merge_after_pooling.add_previous_module(self.conv_merge)
+        conv_merge_after_pooling.add_next_module(new_linear_merge)
+        new_linear_merge.add_previous_module(conv_merge_after_pooling)
+        new_linear_merge.add_next_module(new_linear)
+        new_linear.previous_module = new_linear_merge
+
+        # Assert tensor shapes
+        self.assertEqual(
+            self.conv.tensor_s._shape,
+            (
+                self.conv_in_features + self.conv.use_bias,
+                self.conv_in_features + self.conv.use_bias,
+            ),
+        )
+        self.assertEqual(
+            self.conv.tensor_m._shape,
+            (self.conv_in_features + self.conv.use_bias, self.out_channels),
+        )
+
+        self.assertEqual(self.conv_merge.total_in_features, self.conv_in_features + 1)
+        self.assertEqual(
+            self.conv_merge.tensor_s._shape,
+            (
+                self.out_channels * self.kernel_size[0] * self.kernel_size[1] + 1,
+                self.out_channels * self.kernel_size[0] * self.kernel_size[1] + 1,
+            ),
+        )
+        self.assertEqual(
+            self.conv_merge.previous_tensor_s._shape,
+            (self.conv_merge.total_in_features, self.conv_merge.total_in_features),
+        )
+        self.assertEqual(
+            self.conv_merge.previous_tensor_m._shape,
+            (self.conv_merge.total_in_features, self.out_channels),
+        )
+
+        self.assertEqual(conv_merge_after_pooling.total_in_features, 0)
+        self.assertEqual(
+            conv_merge_after_pooling.tensor_s._shape,
+            (
+                self.out_channels
+                * conv_merge_after_pooling.kernel_size[0]
+                * conv_merge_after_pooling.kernel_size[1]
+                + 1,
+                self.out_channels
+                * conv_merge_after_pooling.kernel_size[0]
+                * conv_merge_after_pooling.kernel_size[1]
+                + 1,
+            ),
+        )
+        self.assertIsNone(conv_merge_after_pooling.previous_tensor_s)
+        self.assertIsNone(conv_merge_after_pooling.previous_tensor_m)
+
+        self.assertEqual(new_linear_merge.total_in_features, 0)
+        self.assertEqual(
+            new_linear_merge.tensor_s._shape,
+            (new_linear_in_features + 1, new_linear_in_features + 1),
+        )
+        self.assertIsNone(new_linear_merge.previous_tensor_s)
+        self.assertIsNone(new_linear_merge.previous_tensor_m)
+
+        self.assertEqual(
+            new_linear.tensor_s._shape,
+            (
+                new_linear_in_features + new_linear.use_bias,
+                new_linear_in_features + new_linear.use_bias,
+            ),
+        )
+        self.assertEqual(
+            new_linear.tensor_m._shape,
+            (new_linear_in_features + new_linear.use_bias, self.out_features),
+        )
+
+        # ---------- Average Pooling ------------
+
+        # Forward and backward pass
+        for layer in self.layers:
+            layer.init_computation()
+
+        x_conv = self.conv(self.x)
+        x_conv_merge = self.conv_merge(x_conv)
+        x_pooling = pooling(x_conv_merge)
+        x_conv_merge_after_pooling = conv_merge_after_pooling(x_pooling)
+        x_flatten = self.flatten(x_conv_merge_after_pooling)
+        x_linear_merge = new_linear_merge(x_flatten)
+        out = new_linear(x_linear_merge)
+
+        self.assertEqual(out.shape, (self.batch_size, self.out_features))
+
+        loss = self.loss_fn(out, self.y)
+        loss.backward()
+
+        for layer in self.layers:
+            layer.update_computation()
+
+        # Compute optimal deltas
+        for layer in self.layers:
+            if isinstance(layer, LinearMergeGrowingModule) or (
+                layer is conv_merge_after_pooling
+            ):
+                with self.assertRaises(AssertionError):
+                    layer.compute_optimal_delta()
+            else:
+                layer.compute_optimal_delta()
+                if isinstance(layer, GrowingModule):
+                    self.assertIsNotNone(layer.optimal_delta_layer)
+
+        # Reset computation
+        for layer in self.layers:
+            layer.reset_computation()
+            layer.delete_update()
+
+        # ---------- Max Pooling ------------
+
+        pooling = torch.nn.MaxPool2d(kernel_size=self.kernel_size)
+
+        # Forward and backward pass
+        for layer in self.layers:
+            layer.init_computation()
+
+        x_conv = self.conv(self.x)
+        x_conv_merge = self.conv_merge(x_conv)
+        x_pooling = pooling(x_conv_merge)
+        x_conv_merge_after_pooling = conv_merge_after_pooling(x_pooling)
+        x_flatten = self.flatten(x_conv_merge_after_pooling)
+        x_linear_merge = new_linear_merge(x_flatten)
+        out = new_linear(x_linear_merge)
+
+        self.assertEqual(out.shape, (self.batch_size, self.out_features))
+
+        loss = self.loss_fn(out, self.y)
+        loss.backward()
+
+        for layer in self.layers:
+            layer.update_computation()
+
+        # Compute optimal deltas
+        for layer in self.layers:
+            if isinstance(layer, LinearMergeGrowingModule) or (
+                layer is conv_merge_after_pooling
+            ):
+                with self.assertRaises(AssertionError):
+                    layer.compute_optimal_delta()
+            else:
+                layer.compute_optimal_delta()
+                if isinstance(layer, GrowingModule):
+                    self.assertIsNotNone(layer.optimal_delta_layer)
+
+        # Reset computation
+        for layer in self.layers:
+            layer.reset_computation()
+            layer.delete_update()
+
+    def test_container_to_container(self):
+        # Create Containers
+        hidden_channels = 10
+        dag1 = GrowingGraphNetwork(
+            in_features=self.in_channels,
+            out_features=hidden_channels,
+            loss_fn=self.loss_fn,
+            input_shape=self.input_shape,
+            layer_type="convolution",
+        )
+        dag2 = GrowingGraphNetwork(
+            in_features=hidden_channels,
+            out_features=self.out_channels,
+            loss_fn=self.loss_fn,
+            input_shape=self.input_shape,
+            layer_type="convolution",
+        )
+
+        start_of_dag1 = dag1.dag.get_node_module(dag1.dag.root)
+        end_of_dag1 = dag1.dag.get_node_module(dag1.dag.end)
+        start_of_dag2 = dag2.dag.get_node_module(dag2.dag.root)
+        end_of_dag2 = dag2.dag.get_node_module(dag2.dag.end)
+        new_linear_in_features = end_of_dag2.output_volume
+
+        new_linear_merge = LinearMergeGrowingModule(
+            in_features=new_linear_in_features, name="new_linear_merge"
+        )
+        new_linear = LinearGrowingModule(
+            in_features=new_linear_in_features,
+            out_features=self.out_features,
+            name="new_linear",
+        )
+        self.layers = [dag1, dag2, new_linear_merge, new_linear]
+
+        # Direct connection between GrowingGraphNetwork and GrowingGraphNetwork
+        end_of_dag1.add_next_module(start_of_dag2)
+        start_of_dag2.add_previous_module(end_of_dag1)
+        end_of_dag2.add_next_module(new_linear_merge)
+        new_linear_merge.add_previous_module(end_of_dag2)
+        new_linear_merge.add_next_module(new_linear)
+        new_linear.previous_module = new_linear_merge
+
+        # Assert tensor shapes
+        self.assertEqual(start_of_dag1.total_in_features, 0)
+        self.assertEqual(
+            start_of_dag1.tensor_s._shape,
+            (
+                self.conv_in_features + self.conv.use_bias,
+                self.conv_in_features + self.conv.use_bias,
+            ),
+        )
+        self.assertIsNone(start_of_dag1.previous_tensor_s)
+        self.assertIsNone(start_of_dag1.previous_tensor_m)
+
+        self.assertEqual(end_of_dag1.total_in_features, self.conv_in_features + 1)
+        self.assertEqual(
+            end_of_dag1.tensor_s._shape,
+            (
+                hidden_channels * self.kernel_size[0] * self.kernel_size[1] + 1,
+                hidden_channels * self.kernel_size[0] * self.kernel_size[1] + 1,
+            ),
+        )
+        self.assertEqual(
+            end_of_dag1.previous_tensor_s._shape,
+            (end_of_dag1.total_in_features, end_of_dag1.total_in_features),
+        )
+        self.assertEqual(
+            end_of_dag1.previous_tensor_m._shape,
+            (end_of_dag1.total_in_features, hidden_channels),
+        )
+
+        self.assertEqual(start_of_dag2.total_in_features, 0)
+        self.assertEqual(
+            start_of_dag2.tensor_s._shape,
+            (
+                hidden_channels * self.kernel_size[0] * self.kernel_size[1] + 1,
+                hidden_channels * self.kernel_size[0] * self.kernel_size[1] + 1,
+            ),
+        )
+        self.assertIsNone(start_of_dag2.previous_tensor_s)
+        self.assertIsNone(start_of_dag2.previous_tensor_m)
+
+        self.assertEqual(
+            end_of_dag2.total_in_features,
+            hidden_channels * self.kernel_size[0] * self.kernel_size[1] + 1,
+        )
+        self.assertEqual(
+            end_of_dag2.tensor_s._shape,
+            (
+                self.out_channels * self.kernel_size[0] * self.kernel_size[1] + 1,
+                self.out_channels * self.kernel_size[0] * self.kernel_size[1] + 1,
+            ),
+        )
+        self.assertEqual(
+            end_of_dag2.previous_tensor_s._shape,
+            (end_of_dag2.total_in_features, end_of_dag2.total_in_features),
+        )
+        self.assertEqual(
+            end_of_dag2.previous_tensor_m._shape,
+            (end_of_dag2.total_in_features, self.out_channels),
+        )
+
+        self.assertEqual(new_linear_merge.total_in_features, 0)
+        self.assertEqual(
+            new_linear_merge.tensor_s._shape,
+            (new_linear_in_features + 1, new_linear_in_features + 1),
+        )
+        self.assertIsNone(new_linear_merge.previous_tensor_s)
+        self.assertIsNone(new_linear_merge.previous_tensor_m)
+
+        self.assertEqual(
+            new_linear.tensor_s._shape,
+            (
+                new_linear_in_features + new_linear.use_bias,
+                new_linear_in_features + new_linear.use_bias,
+            ),
+        )
+        self.assertEqual(
+            new_linear.tensor_m._shape,
+            (new_linear_in_features + new_linear.use_bias, self.out_features),
+        )
+
+        # Forward and backward pass
+        for layer in self.layers:
+            layer.init_computation()
+        start_of_dag2.init_computation()
+
+        x_dag1 = dag1(self.x)
+        x_dag2 = dag2(x_dag1)
+        x_flatten = self.flatten(x_dag2)
+        x_linear_merge = new_linear_merge(x_flatten)
+        out = new_linear(x_linear_merge)
+
+        self.assertEqual(out.shape, (self.batch_size, self.out_features))
+
+        loss = self.loss_fn(out, self.y)
+        loss.backward()
+
+        for layer in self.layers:
+            layer.update_computation()
+
+        # Compute optimal deltas
+        for layer in self.layers:
+            print(type(layer))
+            if isinstance(layer, LinearMergeGrowingModule):
+                with self.assertRaises(AssertionError):
+                    layer.compute_optimal_delta()
+            else:
+                layer.compute_optimal_delta()
+                if isinstance(layer, GrowingModule):
+                    self.assertIsNotNone(layer.optimal_delta_layer)
+                elif isinstance(layer, GrowingGraphNetwork):
+                    for edge_module in layer.dag.get_all_edge_modules():
+                        self.assertIsNotNone(edge_module.optimal_delta_layer)
+
+        # Retrieve bottleneck
+        with torch.no_grad():
+            bottleneck = {  # TODO: rename nodes with dag name as prefix
+                "start": start_of_dag2.projected_v_goal()
+                .clone()
+                .detach(),  # start_of_dag2
+                "end": end_of_dag2.projected_v_goal().clone().detach(),  # start_of_dag2
+            }
+            input_B = {
+                "start": start_of_dag1.activity.clone().detach(),  # start_of_dag1
+                "end": end_of_dag1.activity.clone().detach(),  # end_of_dag1
+            }
+
+        expansion = Expansion(
+            dag=dag2.dag, type="expanded node", expanding_node=dag2.dag.root
+        )
+        actions = [expansion]
+
+        dag2.execute_expansions(
+            actions=actions,
+            bottleneck=bottleneck,
+            input_B=input_B,
+            amplitude_factor=False,
+            evaluate=False,
+            verbose=False,
+        )
+
+        dag2.choose_growth_best_action(
+            options=actions,
+            verbose=False,
+        )
 
         # Reset computation
         for layer in self.layers: